commit ejercicio practico clase 6

+++ calculadora/src/main/java/ar/edu/argentinaprograma/calculadora/Calculadora.java
@@ -2,22 +2,22 @@ package ar.edu.argentinaprograma.calculadora;
 
 public class Calculadora {
 	
-	public static Integer sumar(Integer num1, Integer num2) {
+	public static Double sumar(Double num1, Double num2) {
 		return num1 + num2;
 	}
 	
-	public static Integer restar(Integer num1, Integer num2) {
+	public static Double restar(Double num1, Double num2) {
 		return num1 - num2;
 	}
 	
-	public static Integer multiplicar(Integer num1, Integer num2) {
+	public static Double multiplicar(Double num1, Double num2) {
 		return num1 * num2;
 	}
 	
-	public static Integer dividir(Integer num1, Integer num2) {
+	public static Double dividir(Double num1, Double num2) {
 		return num1 / num2;
 	}
-	public static boolean esMayoroIgualque(Integer num1, Integer num2) {
+	public static boolean esMayoroIgualque(double num1, double num2) {
 		return num1 >= num2;
 	}
